requirements.txt
beautifulsoup4
python-dotenv
ics
requests
google-api-python-client
